+++ tools/killsnoop.py
@@ -46,14 +46,14 @@ bpf_text = """
 #include <linux/sched.h>
 
 struct val_t {
-   u64 pid;
+   u32 pid;
    int sig;
    int tpid;
    char comm[TASK_COMM_LEN];
 };
 
 struct data_t {
-   u64 pid;
+   u32 pid;
    int tpid;
    int sig;
    int ret;
