+++ tools/mysqld_qslower.py
@@ -48,7 +48,7 @@ struct start_t {
 };
 
 struct data_t {
-    u64 pid;
+    u32 pid;
     u64 ts;
     u64 delta;
     char query[QUERY_MAX];
