+++ tools/xfsslower.py
@@ -81,7 +81,7 @@ struct data_t {
     u64 size;
     u64 offset;
     u64 delta_us;
-    u64 pid;
+    u32 pid;
     char task[TASK_COMM_LEN];
     char file[DNAME_INLINE_LEN];
 };
